laravel home

+++ Blogger Website/bloggerWebsite/app/Http/Controllers/userProfileController.php
@@ -10,19 +10,26 @@ use Image;
 
 class userProfileController extends Controller
 {
-    function userProfilePage(){
-        $UserInfo = User::where("email",Auth::user()->email)->get();
-        return view("admin.user.users profile.userProfilePage",[
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
+    function userProfilePage()
+    {
+        $UserInfo = User::where("email", Auth::user()->email)->get();
+        return view("admin.user.users profile.userProfilePage", [
             'UserInfo' => $UserInfo,
 
         ]);
     }
 
 
-    function userProfileUpdate(Request $req){
+    function userProfileUpdate(Request $req)
+    {
         $getImg = $req->photo;
 
-        if($getImg == ""){
+        if ($getImg == "") {
             User::find($req->userId)->update([
                 'name' => $req->name,
                 'email' => $req->email,
@@ -30,26 +37,47 @@ class userProfileController extends Controller
             ]);
 
             return back();
-        }
-
-        else{
-            $delete_pic = public_path('uploads/userProfile/' . Auth::user()->photo);
-            unlink($delete_pic);
-            $extension = $getImg->getClientOriginalExtension();
-            $filename = $req->userId."-".rand().".".$extension;
-
-            Image::make($getImg)->save(public_path("uploads/userProfile/".$filename));
+        } else {
 
 
-            User::find($req->userId)->update([
-                'name' => $req->name,
-                'email' => $req->email,
-                'photo' =>  $filename,
-                'created_at' => Carbon::now(),
-            ]);
+            if (Auth::User()->photo == null) {
 
-            return back()->with("userProfileUpdate","Update Success!!");
 
+                $extension = $getImg->getClientOriginalExtension();
+                $filename = $req->userId . "-" . rand() . "." . $extension;
+
+                Image::make($getImg)->save(public_path("uploads/userProfile/" . $filename));
+
+
+                User::find($req->userId)->update([
+                    'name' => $req->name,
+                    'email' => $req->email,
+                    'photo' =>  $filename,
+                    'created_at' => Carbon::now(),
+                ]);
+
+                return back()->with("userProfileUpdate", "Update Success!!");
+            } else {
+
+                $delete_pic = public_path('uploads/userProfile/' . Auth::user()->photo);
+                unlink($delete_pic);
+
+
+                $extension = $getImg->getClientOriginalExtension();
+                $filename = $req->userId . "-" . rand() . "." . $extension;
+
+                Image::make($getImg)->save(public_path("uploads/userProfile/" . $filename));
+
+
+                User::find($req->userId)->update([
+                    'name' => $req->name,
+                    'email' => $req->email,
+                    'photo' =>  $filename,
+                    'created_at' => Carbon::now(),
+                ]);
+
+                return back()->with("userProfileUpdate", "Update Success!!");
+            }
         }
     }
 }
